requirements.txt
pandas==2.2.3
openpyxl==3.1.5
anvil-uplink==0.4.2
requests==2.32.3
python-dateutil==2.9.0.post0
pdfplumber==0.7.6
